requirements.txt
argparse
blake3
pathlib
yaml
qrcode
pillow
